+++ swanlab/data/callback_local.py
@@ -11,6 +11,7 @@ from swanlab.log import swanlog
 from swanlab.utils.font import FONT
 from swanlab.data.run.main import get_run, SwanLabRunState
 from swanlab.data.run.callback import SwanLabRunCallback, MetricInfo, RuntimeInfo
+from swanlab.env import ROOT
 from datetime import datetime
 import traceback
 import json
@@ -44,6 +45,47 @@ class LocalRunCallback(SwanLabRunCallback):
         else:
             swanlog.info("Error happened while training")
 
+    @staticmethod
+    def _init_logdir(logdir: str = None) -> str:
+        """
+        根据传入的logdir，初始化日志文件夹
+        FIXME shit code
+        """
+        # 如果传入了logdir，则将logdir设置为环境变量，代表日志文件存放的路径
+        if logdir is not None:
+            try:
+                if not isinstance(logdir, str):
+                    raise ValueError("path must be a string")
+                if not os.path.isabs(logdir):
+                    logdir = os.path.abspath(logdir)
+                # 如果创建失败，也是抛出IOError
+                try:
+                    os.makedirs(logdir, exist_ok=True)
+                except Exception as e:
+                    raise IOError(f"create path: {logdir} failed, error: {e}")
+                if not os.access(logdir, os.W_OK):
+                    raise IOError(f"no write permission for path: {logdir}")
+            except ValueError:
+                raise ValueError("logdir must be a str.")
+            except IOError:
+                raise IOError("logdir must be a path and have Write permission.")
+            os.environ[ROOT] = logdir
+        # 如果没有传入logdir，则使用默认的logdir, 即当前工作目录下的swanlog文件夹，但是需要保证目录存在
+        else:
+            logdir = os.environ.get(ROOT) or os.path.join(os.getcwd(), "swanlog")
+            logdir = os.path.abspath(logdir)
+            try:
+                os.makedirs(logdir, exist_ok=True)
+                if not os.access(logdir, os.W_OK):
+                    raise IOError
+            except IOError:
+                raise IOError("logdir must have Write permission.")
+        # 如果logdir是空的，创建.gitignore文件，写入*
+        if not os.listdir(logdir):
+            with open(os.path.join(logdir, ".gitignore"), "w", encoding="utf-8") as f:
+                f.write("*")
+        return logdir
+
     def __str__(self):
         return "SwanLabLocalRunCallback"
 
@@ -66,7 +108,7 @@ class LocalRunCallback(SwanLabRunCallback):
         run.finish() if run.running else swanlog.debug("Duplicate finish, ignore it.")
 
     def on_init(self, proj_name: str, workspace: str, logdir: str = None):
-        pass
+        self._init_logdir(logdir)
 
     def on_run(self):
         swanlog.install(self.settings.console_dir)
