+++ test/unit/cli/pytest_cli_logout.py
